requirements.txt
json==2.0.9
pandas==1.4.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.9